backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Cache
redis==5.0.1
aioredis==2.0.1

# Message Queue
celery==5.3.4
kombu==5.3.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2

# WebSockets
websockets==12.0
socketio==0.0.1

# File handling
python-magic==0.4.27
aiofiles==23.2.1
Pillow==10.1.0

# Search and indexing
elasticsearch==8.11.0
whoosh==2.7.4

# PDF and Excel export
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9

# Email
aiosmtplib==3.0.1
emails==0.6.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# GraphQL (alternative API)
strawberry-graphql==0.213.0

# Operational Transformation
ot==0.0.1

# Internationalization
babel==2.13.1
python-i18n==0.3.9
